pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikolai Trusov",email = "ntrusov@umschool.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "dishka (>=1.4.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "pytest-cov (>=6.0.0,<7.0.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "adaptix (>=3.0.0b9,<4.0.0)",
    "boto3 (>=1.37.17,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)"

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
